separate out Plugin.GHCSpecter.Comm and Console module
