+++ plugin/src/Plugin/GHCSpecter.hs
@@ -9,30 +9,22 @@ module Plugin.GHCSpecter
   )
 where
 
-import Control.Concurrent (forkIO, forkOS, killThread)
+import Control.Concurrent (forkOS)
 import Control.Concurrent.STM
   ( atomically,
     modifyTVar',
     readTVar,
-    retry,
-    writeTVar,
   )
-import Control.Concurrent.STM qualified as STM
-import Control.Exception qualified as E
-import Control.Monad (forever, void, when)
+import Control.Monad (void, when)
 import Control.Monad.IO.Class (liftIO)
 import Data.Foldable (for_)
-import Data.Foldable qualified as F
 import Data.IORef (IORef, newIORef, readIORef, writeIORef)
 import Data.List qualified as L
 import Data.Map (Map)
 import Data.Map qualified as M
-import Data.Sequence ((|>))
-import Data.Sequence qualified as Seq
 import Data.Set (Set)
 import Data.Set qualified as S
 import Data.Text qualified as T
-import Data.Text.IO qualified as TIO
 import Data.Time.Clock (UTCTime, getCurrentTime)
 import GHC.Driver.Env (Hsc, HscEnv (..))
 import GHC.Driver.Flags (GeneralFlag (Opt_WriteHie))
@@ -69,32 +61,17 @@ import GHCSpecter.Channel.Common.Types
   ( DriverId (..),
     type ModuleName,
   )
-import GHCSpecter.Channel.Inbound.Types
-  ( ConsoleRequest (..),
-    Request (..),
-    SessionRequest (..),
-  )
 import GHCSpecter.Channel.Outbound.Types
   ( BreakpointLoc (..),
     ChanMessage (..),
-    ChanMessageBox (..),
     HsSourceInfo (..),
     SessionInfo (..),
     Timer (..),
     TimerTag (..),
   )
-import GHCSpecter.Comm
-  ( receiveObject,
-    runClient,
-    sendObject,
-  )
-import GHCSpecter.Config
-  ( Config (..),
-    defaultGhcSpecterConfigFile,
-    loadConfig,
-  )
 import GHCSpecter.Util.GHC (showPpr)
-import Network.Socket (Socket)
+import Plugin.GHCSpecter.Comm (queueMessage, runMessageQueue)
+import Plugin.GHCSpecter.Console (breakPoint)
 import Plugin.GHCSpecter.Types
   ( MsgQueue (..),
     PluginSession (..),
@@ -108,112 +85,8 @@ import Plugin.GHCSpecter.Util
     getModuleName,
     mkModuleNameMap,
   )
-import System.Directory (canonicalizePath, doesFileExist)
+import System.Directory (canonicalizePath)
 import System.Process (getCurrentPid)
-
-runMessageQueue :: [CommandLineOption] -> MsgQueue -> IO ()
-runMessageQueue opts queue = do
-  mipcfile <-
-    case opts of
-      ipcfile : _ -> pure (Just ipcfile)
-      [] -> do
-        ecfg <- loadConfig defaultGhcSpecterConfigFile
-        case ecfg of
-          Left _ -> pure Nothing
-          Right cfg -> pure (Just (configSocket cfg))
-  for_ mipcfile $ \ipcfile -> do
-    socketExists <- doesFileExist ipcfile
-    when socketExists $
-      runClient ipcfile $ \sock -> do
-        _ <- forkIO $ receiver sock
-        sender sock
-  where
-    sender :: Socket -> IO ()
-    sender sock = forever $ do
-      msgs <- atomically $ do
-        queued <- readTVar (msgSenderQueue queue)
-        if Seq.null queued
-          then retry
-          else do
-            writeTVar (msgSenderQueue queue) Seq.empty
-            pure queued
-      let msgList = F.toList msgs
-      msgList `seq` sendObject sock msgList
-    receiver :: Socket -> IO ()
-    receiver sock = forever $ do
-      putStrLn "################"
-      putStrLn "receiver started"
-      putStrLn "################"
-      msg :: Request <- receiveObject sock
-      putStrLn "################"
-      putStrLn $ "message received: " ++ show msg
-      putStrLn "################"
-      atomically $
-        case msg of
-          SessionReq sreq ->
-            modifyTVar' sessionRef $ \s ->
-              let isPaused
-                    | sreq == Pause = True
-                    | otherwise = False
-                  sinfo = psSessionInfo s
-                  sinfo' = sinfo {sessionIsPaused = isPaused}
-               in s {psSessionInfo = sinfo'}
-          ConsoleReq drvId' creq ->
-            writeTVar (msgReceiverQueue queue) (Just (drvId', creq))
-
-queueMessage :: MsgQueue -> ChanMessage a -> IO ()
-queueMessage queue !msg =
-  atomically $
-    modifyTVar' (msgSenderQueue queue) (|> CMBox msg)
-
-breakPoint :: MsgQueue -> DriverId -> BreakpointLoc -> IO ()
-breakPoint queue drvId loc = do
-  tid <- forkIO $ sessionInPause queue drvId loc
-  atomically $ do
-    psess <- readTVar sessionRef
-    let sinfo = psSessionInfo psess
-        isSessionPaused = sessionIsPaused sinfo
-        isDriverInStep = maybe False (== drvId) $ psDriverInStep psess
-    -- block until the session is resumed.
-    STM.check (not isSessionPaused || isDriverInStep)
-    when (isDriverInStep && isSessionPaused) $ do
-      let psess' = psess {psDriverInStep = Nothing}
-      writeTVar sessionRef psess'
-  killThread tid
-
-consoleAction :: MsgQueue -> DriverId -> IO ()
-consoleAction queue drvId = do
-  let rQ = msgReceiverQueue queue
-  req <-
-    atomically $ do
-      mreq <- readTVar rQ
-      case mreq of
-        Nothing -> retry
-        Just (drvId', req) ->
-          if drvId == drvId'
-            then do
-              writeTVar rQ Nothing
-              pure req
-            else retry
-  case req of
-    Ping msg -> do
-      TIO.putStrLn $ "ping: " <> msg
-      let pongMsg = "pong: " <> msg
-      queueMessage queue (CMConsole drvId pongMsg)
-    NextBreakpoint -> do
-      putStrLn "NextBreakpoint"
-      atomically $
-        modifyTVar' sessionRef $ \psess -> psess {psDriverInStep = Just drvId}
-
-sessionInPause :: MsgQueue -> DriverId -> BreakpointLoc -> IO ()
-sessionInPause queue drvId loc = do
-  atomically $ do
-    isPaused <- sessionIsPaused . psSessionInfo <$> readTVar sessionRef
-    -- prodceed when the session is paused.
-    STM.check isPaused
-  queueMessage queue (CMPaused drvId (Just loc))
-  (forever $ consoleAction queue drvId)
-    `E.catch` (\(_e :: E.AsyncException) -> queueMessage queue (CMPaused drvId Nothing))
 
 -- | Called only once for sending session information
 startSession ::
